Updated performAccessCheck doc
