+++ docs/permissions/custom_policies.md
@@ -251,5 +251,5 @@ For example, `translations/ibexa_content_forms_policies.en.yaml`:
 Check if current user has this custom limitation set to true from a custom controller:
 
 ```php
-[[= include_file('code_samples/back_office/limitation/src/Controller/CustomController.php') =]]
+[[= include_file('code_samples/back_office/limitation/src/Controller/CustomLimitationController.php') =]]
 ```
